DOCS: rastreabilidade melhorada II
Let's hope this is the last menial and very manual work I must do in this project. This required no critical thinking whatsoever, just a series of

open this document
verify it is the same as in the general document
oh foxtrot, it ain't.
dance around like a little miserable rat
done

in series

mindnumbing! Not. Great! NOT GREAT!!